+++ communicator/winrm/file_manager.go
@@ -1,4 +1,4 @@
-package main
+package winrm
 
 import (
 	"encoding/base64"
